+++ matrix-sdk/src/androidTest/java/org/matrix/androidsdk/CryptoTestHelper.java
@@ -132,8 +132,8 @@ public class CryptoTestHelper {
 
         IMXStore store = new MXFileStore(hs, context);
 
-        MXSession mxSession = new MXSession(hs, new MXDataHandler(store, credentials), context);
-
+        MXSession mxSession = new MXSession.Builder(hs, new MXDataHandler(store, credentials), context)
+                .build();
 
         if (!startSession) {
             return mxSession;
@@ -194,7 +194,8 @@ public class CryptoTestHelper {
 
         IMXStore store = new MXFileStore(hs, context);
 
-        MXSession mxSession = new MXSession(hs, new MXDataHandler(store, credentials), context);
+        MXSession mxSession = new MXSession.Builder(hs, new MXDataHandler(store, credentials), context)
+                .build();
 
         mxSession.enableCryptoWhenStarting();