+++ CHANGES.rst
@@ -1,3 +1,17 @@
+Changes to Matrix Android SDK in 0.9.9 (2018-08-30)
+=======================================================
+
+Improvements:
+ - Clear unreachable Url when clearing media cache (vector-im/riot-android#2479)
+ - "In reply to" is not clickable on Riot Android yet. Make it a plain text (vector-im/riot-android#2469)
+
+Bugfix:
+ - Removing room from 'low priority' or 'favorite' does not work (vector-im/riot-android#2526)
+ - MatrixError mResourceLimitExceededError is now managed in MxDataHandler (vector-im/riot-android#2547)
+
+API Change:
+ - MxSession constructor is now private. Please use MxSession.Builder() to create a MxSession
+
 Changes to Matrix Android SDK in 0.9.8 (2018-08-27)
 =======================================================